Revert "feat: change routing (#111)"
This reverts commit dc4d8db836492b235c41460cd7b1467d9c94ebe6.
