--- src/app/(home)/@leaderboardTable/not-found.tsx
@@ -1,3 +1,0 @@
-import { LeaderboardTableLoading } from "@/app/_components/@leaderboardTable";
-
-export default LeaderboardTableLoading;
