--- src/app/(home)/@leaderboardTable/page.tsx
@@ -1,3 +1,0 @@
-import { LeaderboardTable } from "@/app/_components/@leaderboardTable";
-
-export default LeaderboardTable;
