--- src/app/(home)/@searchBuilder/page.tsx
@@ -1,3 +1,0 @@
-import { SearchBuilder } from "@/app/_components/@searchBuilder";
-
-export default SearchBuilder;
