+++ src/app/_components/@searchBuilder/index.ts
@@ -1,2 +1,1 @@
 export { default as SearchBuilder } from "./page";
-export { default as SearchBuilderLoading } from "./loading";
