+++ src/app/airdrop/@bossPointsCard/component.tsx
@@ -33,7 +33,3 @@ export const BossPointsCardComponent: FunctionComponent<
     </BlockyCard>
   );
 };
-
-export const BossPointsCardLoadingComponent = () => {
-  return <BossPointsCardComponent points={"--.---"} />;
-};
