--- src/app/airdrop/@bossPointsCard/not-found.tsx
@@ -1,3 +1,0 @@
-import { BossPointsCardLoadingComponent } from "./component";
-
-export default BossPointsCardLoadingComponent;
