--- src/app/airdrop/@bossTokensCard/not-found.tsx
@@ -1,3 +1,0 @@
-import { BossTokensCardLoading } from "./component";
-
-export default BossTokensCardLoading;
