--- src/app/airdrop/@bossTokensCard/page.tsx
@@ -1,9 +1,0 @@
-import { notFound } from "next/navigation";
-import { getCurrentUser } from "@/app/_api/get-user";
-import { BossTokensCardComponent } from "./component";
-
-export default async function BossTokensCard() {
-  const user = await getCurrentUser();
-  if (!user) return notFound();
-  return <BossTokensCardComponent tokens={user.boss_token_balance} />;
-}
