--- src/app/airdrop/@builderScoreCard/loading.tsx
@@ -1,3 +1,0 @@
-import { BuilderScoreCardLoading } from "./component";
-
-export default BuilderScoreCardLoading;
