--- src/app/airdrop/@dailyBudgetCard/default.tsx
@@ -1,3 +1,0 @@
-import { DailyBudgetCardLoading } from "./component";
-
-export default DailyBudgetCardLoading;
