+++ src/app/airdrop/@dailyBudgetCard/loading.tsx
@@ -1,3 +1,5 @@
-import { DailyBudgetCardLoading } from "./component";
+import { DailyBudgetCardComponent } from "./component";
 
-export default DailyBudgetCardLoading;
+export default function DailyBudgetCardLoading() {
+  return <DailyBudgetCardComponent loading />;
+}
