--- src/app/airdrop/@dailyStreakCard/not-found.tsx
@@ -1,3 +1,0 @@
-import { DailyStreakCardLoading } from "./component";
-
-export default DailyStreakCardLoading;
