--- src/app/airdrop/@howToPlay/page.tsx
@@ -1,3 +1,0 @@
-import { HowToPlay } from "@/app/_components/@howToPlay";
-
-export default HowToPlay;
