+++ src/app/airdrop/@myNominationsTable/component.tsx
@@ -49,7 +49,3 @@ export const MyNominationsTableComponent: FunctionComponent<
     </Table>
   </Sheet>
 );
-
-export const MyNominationsTableLoading = () => (
-  <MyNominationsTableComponent loading />
-);
