--- src/app/airdrop/@myNominationsTable/not-found.tsx
@@ -1,3 +1,0 @@
-import { MyNominationsTableLoading } from "./component";
-
-export default MyNominationsTableLoading;
