--- src/app/airdrop/@searchBuilder/default.tsx
@@ -1,3 +1,0 @@
-import { SearchBuilderLoading } from "@/app/_components/@searchBuilder";
-
-export default SearchBuilderLoading;
